requirements.txt
# Project dependencies
pandas
python-dotenv
openpyxl
xlsxwriter